+++ dwds/lib/src/debugging/instance.dart
@@ -1,6 +1,8 @@
 // Copyright (c) 2019, the Dart project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
-// BSD-style license that can be found in the LICENSE file.import 'dart:async';
+// BSD-style license that can be found in the LICENSE file.
+
+import 'dart:math';
 
 import 'package:webkit_inspection_protocol/webkit_inspection_protocol.dart';
 
@@ -9,82 +11,119 @@ import '../utilities/domain.dart';
 import '../utilities/objects.dart';
 import '../utilities/shared.dart';
 import '../utilities/wrapped_service.dart';
-import 'debugger.dart';
+import 'classes.dart';
 import 'inspector.dart';
 import 'metadata.dart';
-import 'remote_debugger.dart';
 
-/// Creates an [InstanceRef] for a primitive [RemoteObject].
-InstanceRef _primitiveInstance(String kind, RemoteObject remoteObject) {
-  var classRef = ClassRef(
-      // TODO(grouma) - is this ID correct?
-      id: 'dart:core:${remoteObject?.type}',
-      name: kind);
-  return InstanceRef(
-      kind: kind, classRef: classRef, id: dartIdFor(remoteObject?.value))
-    ..valueAsString = '${remoteObject?.value}';
-}
-
-/// A hard-coded ClassRef for the String class.
-final _classRefForString =
-    ClassRef(id: 'dart:core:String', name: InstanceKind.kString);
-
-/// A hard-coded ClassRef for the Closure class.
-// TODO(grouma) - orgnaize our static classRefs better.
-final _classRefForClosure = ClassRef(name: 'Closure', id: createId());
-
-/// A hard-coded ClassRef for a (non-existent) class called Unknown.
-final _classRefForUnknown = ClassRef(name: 'Unknown', id: createId());
+/// The maximum length of String which will be returned in an InstanceRef.
+///
+/// Anything longer will be truncated.
+const _stringTruncationLimit = 128;
 
 /// Contains a set of methods for getting [Instance]s and [InstanceRef]s.
 class InstanceHelper extends Domain {
-  final Debugger _debugger;
-  final RemoteDebugger _remoteDebugger;
+  InstanceHelper(AppInspector Function() provider) : super(provider);
 
-  InstanceHelper(
-      this._debugger, this._remoteDebugger, AppInspector Function() provider)
-      : super(provider);
+  /// Creates an [InstanceRef] for a primitive [RemoteObject].
+  InstanceRef _primitiveInstanceRef(String kind, RemoteObject remoteObject) {
+    var classRef = classRefFor('dart:core', kind);
+    return InstanceRef(
+        kind: kind, classRef: classRef, id: dartIdFor(remoteObject?.value))
+      ..valueAsString = '${remoteObject?.value}';
+  }
 
-  Future<Instance> _stringInstanceFor(RemoteObject remoteObject) async {
-    var actualString = stringFromDartId(remoteObject.objectId);
+  /// Creates an [Instance] for a primitive [RemoteObject].
+  Instance _primitiveInstance(String kind, RemoteObject remote) {
+    if (remote?.objectId == null) return null;
     return Instance(
-        kind: InstanceKind.kString,
-        classRef: _classRefForString,
-        id: createId())
-      ..valueAsString = actualString
-      ..length = actualString.length;
+        id: remote.objectId,
+        kind: kind,
+        classRef: classRefFor('dart:core', kind))
+      ..valueAsString = '${remote.value}';
+  }
+
+  Instance _stringInstanceFor(
+      RemoteObject remoteObject, int offset, int count) {
+    // TODO(#777) Consider a way of not passing the whole string around (in the
+    // ID) in order to find a substring.
+    var fullString = stringFromDartId(remoteObject.objectId);
+    var preview = fullString;
+    var truncated = false;
+    if (offset != null || count != null) {
+      truncated = true;
+      var start = offset ?? 0;
+      var end = count == null ? null : min(start + count, fullString.length);
+      preview = fullString.substring(start, end);
+    }
+    return Instance(
+        kind: InstanceKind.kString, classRef: classRefForString, id: createId())
+      ..valueAsString = preview
+      ..valueAsStringIsTruncated = truncated
+      ..length = fullString.length
+      ..count = (truncated ? preview.length : null)
+      ..offset = (truncated ? offset : null);
   }
 
   Future<Instance> _closureInstanceFor(RemoteObject remoteObject) async {
     var result = Instance(
         kind: InstanceKind.kClosure,
         id: remoteObject.objectId,
-        classRef: _classRefForClosure);
+        classRef: classRefForClosure);
     return result;
   }
 
   /// Create an [Instance] for the given [remoteObject].
   ///
   /// Does a remote eval to get instance information. Returns null if there
-  /// isn't a corresponding instance.
-  Future<Instance> instanceFor(RemoteObject remoteObject) async {
-    if (isStringId(remoteObject.objectId)) {
-      return _stringInstanceFor(remoteObject);
+  /// isn't a corresponding instance. For enumerable objects, [offset] and
+  /// [count] allow retrieving a sub-range of properties.
+  Future<Instance> instanceFor(RemoteObject remoteObject,
+      {int offset, int count}) async {
+    var primitive = _primitiveInstanceOrNull(remoteObject, offset, count);
+    if (primitive != null) {
+      return primitive;
     }
+    if (isStringId(remoteObject.objectId)) {
+      return _stringInstanceFor(remoteObject, offset, count);
+    }
+
     var metaData = await ClassMetaData.metaDataFor(
-        _remoteDebugger, remoteObject, inspector);
+        inspector.remoteDebugger, remoteObject, inspector);
     var classRef = metaData.classRef;
     if (metaData.jsName == 'Function') {
       return _closureInstanceFor(remoteObject);
     }
-    var properties = await _debugger.getProperties(remoteObject.objectId);
-    if (metaData.jsName == 'JSArray') {
-      return await _listInstanceFor(classRef, remoteObject, properties);
-    } else if (metaData.jsName == 'LinkedMap' ||
-        metaData.jsName == 'IdentityMap') {
-      return await _mapInstanceFor(classRef, remoteObject, properties);
+    var properties = await inspector.debugger.getProperties(
+        remoteObject.objectId,
+        offset: offset,
+        count: count,
+        length: metaData.length);
+    if (metaData.isSystemList) {
+      return await _listInstanceFor(
+          classRef, remoteObject, properties, offset, count);
+    } else if (metaData.isSystemMap) {
+      return await _mapInstanceFor(
+          classRef, remoteObject, properties, offset, count);
     } else {
       return await _plainInstanceFor(classRef, remoteObject, properties);
+    }
+  }
+
+  /// If [remoteObject] represents a primitive, return an [Instance] for it,
+  /// otherwise return null.
+  Instance _primitiveInstanceOrNull(
+      RemoteObject remoteObject, int offset, int count) {
+    switch (remoteObject?.type ?? 'undefined') {
+      case 'string':
+        return _stringInstanceFor(remoteObject, offset, count);
+      case 'number':
+        return _primitiveInstance(InstanceKind.kDouble, remoteObject);
+      case 'boolean':
+        return _primitiveInstance(InstanceKind.kBool, remoteObject);
+      case 'undefined':
+        return _primitiveInstance(InstanceKind.kNull, remoteObject);
+      default:
+        return null;
     }
   }
 
@@ -126,55 +165,82 @@ class InstanceHelper extends Domain {
   }
 
   /// The associations for a Dart Map or IdentityMap.
-  Future<List<MapAssociation>> _mapAssociations(RemoteObject map) async {
+  Future<List<MapAssociation>> _mapAssociations(
+      RemoteObject map, int offset, int count) async {
+    // We do this in in awkward way because we want the keys and values, but we
+    // can't return things by value or some Dart objects will come back as
+    // values that we need to be RemoteObject, e.g. a List of int.
     var expression = '''
       function() {
         var sdkUtils = $loadModule('dart_sdk').dart;
         var entries = sdkUtils.dloadRepl(this, "entries");
         entries = sdkUtils.dsendRepl(entries, "toList", []);
-        function asKeyValue(entry) {
-          return {
-            key: sdkUtils.dloadRepl(entry, "key"),
-            value: sdkUtils.dloadRepl(entry, "value")
-          }
+        function asKey(entry) {
+          return sdkUtils.dloadRepl(entry, "key");
         }
-        return entries.map(asKeyValue);
+        function asValue(entry) {
+          return sdkUtils.dloadRepl(entry, "value");
+        }
+        return {
+          keys: entries.map(asKey),
+          values: entries.map(asValue)
+        };
       }
     ''';
-    var keysAndValues = await inspector.jsCallFunctionOn(map, expression, [],
-        returnByValue: true);
+    var keysAndValues = await inspector.jsCallFunctionOn(map, expression, []);
+    var keys = await inspector.loadField(keysAndValues, 'keys');
+    var values = await inspector.loadField(keysAndValues, 'values');
+    var keysInstance = await instanceFor(keys, offset: offset, count: count);
+    var valuesInstance =
+        await instanceFor(values, offset: offset, count: count);
     var associations = <MapAssociation>[];
-    for (var each in keysAndValues.value as List) {
+    Map.fromIterables(keysInstance.elements, valuesInstance.elements)
+        .forEach((key, value) {
       associations.add(MapAssociation()
-        ..key = await instanceRefFor(each['key'])
-        ..value = await instanceRefFor(each['value']));
-    }
+        ..key = key
+        ..value = value);
+    });
     return associations;
   }
 
   /// Create a Map instance with class [classRef] from [remoteObject].
-  Future<Instance> _mapInstanceFor(
-      ClassRef classRef, RemoteObject remoteObject, List<Property> _) async {
+  Future<Instance> _mapInstanceFor(ClassRef classRef, RemoteObject remoteObject,
+      List<Property> _, int offset, int count) async {
     // Maps are complicated, do an eval to get keys and values.
-    var associations = await _mapAssociations(remoteObject);
+    var associations = await _mapAssociations(remoteObject, offset, count);
+    var length = (offset == null && count == null)
+        ? associations.length
+        : (await instanceRefFor(remoteObject)).length;
     return Instance(
         kind: InstanceKind.kMap, id: remoteObject.objectId, classRef: classRef)
-      ..length = associations.length
+      ..length = length
+      ..offset = offset
+      ..count = (associations.length == length) ? null : associations.length
       ..associations = associations;
   }
 
   /// Create a List instance of [classRef] from [remoteObject] with the JS
   /// properties [properties].
-  Future<Instance> _listInstanceFor(ClassRef classRef,
-      RemoteObject remoteObject, List<Property> properties) async {
-    var length = _lengthOf(properties);
-    var indexed = properties.sublist(0, length);
+  Future<Instance> _listInstanceFor(
+      ClassRef classRef,
+      RemoteObject remoteObject,
+      List<Property> properties,
+      int offset,
+      int count) async {
+    var numberOfProperties = _lengthOf(properties);
+    var length = (offset == null && count == null)
+        ? numberOfProperties
+        : (await instanceRefFor(remoteObject)).length;
+    var indexed =
+        properties.sublist(0, min(count ?? length, numberOfProperties));
     var fields = await Future.wait(indexed
         .map((property) async => await _instanceRefForRemote(property.value)));
     return Instance(
         kind: InstanceKind.kList, id: remoteObject.objectId, classRef: classRef)
       ..length = length
-      ..elements = fields;
+      ..elements = fields
+      ..offset = offset
+      ..count = (numberOfProperties == length) ? null : numberOfProperties;
   }
 
   /// Return the value of the length attribute from [properties], if present.
@@ -205,12 +271,12 @@ class InstanceHelper extends Domain {
       const sdk = $loadModule("dart_sdk");
       const sdk_utils = sdk.dart;
       const fields = sdk_utils.getFields(sdk_utils.getType(this)) || [];
-      if (!fields && (dart_sdk._interceptors.JSArray.is(this) || 
+      if (!fields && (dart_sdk._interceptors.JSArray.is(this) ||
           dart_sdk._js_helper.InternalMap.is(this))) {
         // Trim off the 'length' property.
         const fields = allJsProperties.slice(0, allJsProperties.length -1);
         return fields.join(',');
-      } 
+      }
       const privateFields = sdk_utils.getOwnPropertySymbols(fields);
       const nonSymbolNames = privateFields.map(sym => sym.description);
       const publicFieldNames = sdk_utils.getOwnPropertyNames(fields);
@@ -251,37 +317,43 @@ class InstanceHelper extends Domain {
   Future<InstanceRef> _instanceRefForRemote(RemoteObject remoteObject) async {
     // If we have a null result, treat it as a reference to null.
     if (remoteObject == null) {
-      return _primitiveInstance(InstanceKind.kNull, remoteObject);
+      return _primitiveInstanceRef(InstanceKind.kNull, remoteObject);
     }
     switch (remoteObject.type) {
       case 'string':
+        var stringValue = remoteObject.value as String;
+        var truncated = stringValue.length > _stringTruncationLimit;
+        var result = truncated
+            ? stringValue.substring(0, _stringTruncationLimit)
+            : stringValue;
         return InstanceRef(
             id: dartIdFor(remoteObject.value),
-            classRef: _classRefForString,
+            classRef: classRefForString,
             kind: InstanceKind.kString)
-          ..valueAsString = remoteObject.value as String;
+          ..valueAsString = result
+          ..valueAsStringIsTruncated = truncated
+          ..length = stringValue.length;
       case 'number':
-        return _primitiveInstance(InstanceKind.kDouble, remoteObject);
+        return _primitiveInstanceRef(InstanceKind.kDouble, remoteObject);
       case 'boolean':
-        return _primitiveInstance(InstanceKind.kBool, remoteObject);
+        return _primitiveInstanceRef(InstanceKind.kBool, remoteObject);
       case 'undefined':
-        return _primitiveInstance(InstanceKind.kNull, remoteObject);
+        return _primitiveInstanceRef(InstanceKind.kNull, remoteObject);
       case 'object':
         if (remoteObject.objectId == null) {
-          return _primitiveInstance(InstanceKind.kNull, remoteObject);
+          return _primitiveInstanceRef(InstanceKind.kNull, remoteObject);
         }
         var metaData = await ClassMetaData.metaDataFor(
-            _remoteDebugger, remoteObject, inspector);
+            inspector.remoteDebugger, remoteObject, inspector);
         if (metaData == null) return null;
-        if (metaData.jsName == 'JSArray') {
+        if (metaData.isSystemList) {
           return InstanceRef(
               kind: InstanceKind.kList,
               id: remoteObject.objectId,
               classRef: metaData.classRef)
             ..length = metaData.length;
         }
-        if (metaData.jsName == 'LinkedMap' ||
-            metaData.jsName == 'IdentityMap') {
+        if (metaData.isSystemMap) {
           return InstanceRef(
               kind: InstanceKind.kMap,
               id: remoteObject.objectId,
@@ -293,18 +365,18 @@ class InstanceHelper extends Domain {
             id: remoteObject.objectId,
             classRef: metaData.classRef);
       case 'function':
-        var functionMetaData =
-            await FunctionMetaData.metaDataFor(_remoteDebugger, remoteObject);
+        var functionMetaData = await FunctionMetaData.metaDataFor(
+            inspector.remoteDebugger, remoteObject);
         return InstanceRef(
             kind: InstanceKind.kClosure,
             id: remoteObject.objectId,
-            classRef: _classRefForClosure)
+            classRef: classRefForClosure)
           // TODO(grouma) - fill this in properly.
           ..closureFunction = FuncRef(
               name: functionMetaData.name,
               id: createId(),
               // TODO(alanknight): The right ClassRef
-              owner: _classRefForUnknown,
+              owner: classRefForUnknown,
               isConst: false,
               isStatic: false)
           ..closureContext = (ContextRef()..length = 0);